refactor: 불필요한 casting 제거, test 에 만 사용되는 생성자 제거

+++ src/main/java/lotto/domain/Lotto.java
@@ -13,7 +13,7 @@ public class Lotto {
         this.number = number;
     }
 
-    public int getMatchedNumberCount(NumberMatcher numberMatcher) {
+    public long getMatchedNumberCount(NumberMatcher numberMatcher) {
         return this.number.getMatchedNumberCount(numberMatcher);
     }
